backend/requirements.txt
# requirements.txt
# ============================================================================
# BARRIOS A2I WEBSITE ASSISTANT v2.0 — PYTHON DEPENDENCIES
# ============================================================================

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# LangChain/LangGraph
langgraph>=0.0.40
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Database (optional - for production)
supabase>=2.3.0
asyncpg>=0.29.0

# Observability
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
prometheus-client>=0.19.0

# Utilities
python-dotenv>=1.0.0
orjson>=3.9.0

# Google Drive Storage
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
